+++ packages/contracts/contracts/test/plugin-mock/UUPSUpgradeable/PluginSetupV1MockBad.sol
@@ -13,7 +13,7 @@ contract PluginSetupV1MockBad is PluginSetup {
 
     uint256 public constant PLUGIN_INIT_NUMBER = 15;
 
-    address private noOracle;
+    address private constant NO_ORACLE = address(0);
 
     constructor() {
         // User the plugin as helper for testing puposes.
@@ -67,7 +67,7 @@ contract PluginSetupV1MockBad is PluginSetup {
             PermissionLib.Operation.Grant,
             _dao,
             plugin,
-            noOracle,
+            NO_ORACLE,
             keccak256("EXECUTE_PERMISSION")
         );
 
@@ -76,7 +76,7 @@ contract PluginSetupV1MockBad is PluginSetup {
                 PermissionLib.Operation.Grant,
                 _dao,
                 plugin,
-                noOracle,
+                NO_ORACLE,
                 keccak256("BAD_PERMISSION")
             );
         }
@@ -99,7 +99,7 @@ contract PluginSetupV1MockBad is PluginSetup {
             PermissionLib.Operation.Revoke,
             _dao,
             _plugin,
-            noOracle,
+            NO_ORACLE,
             keccak256("EXECUTE_PERMISSION")
         );
 
@@ -108,7 +108,7 @@ contract PluginSetupV1MockBad is PluginSetup {
                 PermissionLib.Operation.Grant,
                 _plugin,
                 _activeHelpers[0],
-                noOracle,
+                NO_ORACLE,
                 keccak256("BAD_PERMISSION")
             );
         }